refactor move up ok
