+++ Makefile
@@ -6,7 +6,7 @@
 #    By: wwalas- <wwallas-@student.42sp.org.br>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/11/29 16:09:22 by wwallas-          #+#    #+#              #
-#    Updated: 2023/01/23 16:00:49 by wwalas-          ###   ########.fr        #
+#    Updated: 2023/01/24 12:06:52 by wwalas-          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -22,9 +22,9 @@ INCLUDE		=	-I./libft	\
 
 SOURCES		=	verify_extension.c open_file.c valid_map.c valid_chars.c valid_chars_ults.c verify_exit.c	\
 				alloc_map.c draw_map.c init_data.c init_img.c init_mlx.c init_windows.c draw_map_debug.c	\
-				print_larger_pixel.c save_x.c delet_map.c filter_keyboard.c update_dir.c  update_position.c \
-				updates.c start_program.c control.c raycast.c calculetePixel.c positionMap.c radiusCalculation.c		\
-				screen.c
+				print_larger_pixel.c save_x.c delet_map.c filter_keyboard.c update_dir.c updates.c \
+				start_program.c control.c raycast.c calculetePixel.c positionMap.c radiusCalculation.c		\
+				screen.c is_valid_position.c move_up.c move_down.c
 
 OBJS_DIR	=	object
 OBJECTS		=	$(patsubst %.c, $(OBJS_DIR)/%.o, $(SOURCES))
