+++ include/cub3D.h
@@ -6,7 +6,7 @@
 /*   By: wwalas- <wwallas-@student.42sp.org.br>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/30 13:32:06 by wwallas-          #+#    #+#             */
-/*   Updated: 2023/01/23 16:29:36 by wwalas-          ###   ########.fr       */
+/*   Updated: 2023/01/24 12:05:41 by wwalas-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -110,7 +110,7 @@ void		map_move_right(t_map *map);
 void		map_move_left(t_map *map);
 void		map_move_down(t_map *map);
 void		update_p_char(t_map *map, int x, int y);
-void		update_p_player(t_map *map, int x, int y);
+void		update_p_player(t_map *map, double position_y, double position_x);
 
 void		draw_map_debug(char **map);
 void		testes(t_data *data);
@@ -134,5 +134,7 @@ void	lengthRayNext_x_y(t_raycast *itens, t_data *data);
 void	jump_next_square_and_verify_hit_wall(t_raycast * itens, t_data *data, int hit);
 void	calculatePixel(t_raycast *itens, t_data *data);
 
+
+int		is_valid_poosition(t_map *map, double x, double y);
 
 #endif
