+++ sources/program/is_valid_position.c
@@ -1,24 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   updates.c                                          :+:      :+:    :+:   */
+/*   is_valid_position.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: wwalas- <wwallas-@student.42sp.org.br>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 18:34:25 by wwallas-          #+#    #+#             */
-/*   Updated: 2023/01/24 11:45:48 by wwalas-          ###   ########.fr       */
+/*   Updated: 2023/01/24 11:49:23 by wwalas-          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <cub3D.h>
 
-void	update_p_char(t_map *map, int x, int y)
+int	is_valid_poosition(t_map *map, double x, double y)
 {
-	map->map[y][x] = 'W';
-}
-
-void	update_p_player(t_map *map, double position_y, double position_x)
-{
-	map->player[P_X] = position_x;
-	map->player[P_Y] = position_y;
+	if (map->map[(int)y][(int)x] == '1')
+		return (FALSE);
+	return (TRUE);
 }
